+++ lib/interactify/configuration.rb
@@ -11,5 +11,13 @@ module Interactify
     def fallback
       Rails.root / "app" if Interactify.railties?
     end
+
+    def trigger_definition_error(error)
+      @on_definition_error&.call(error)
+    end
+
+    def on_definition_error(handler = nil, &block)
+      @on_definition_error = block_given? ? block : handler
+    end
   end
 end