+++ lib/interactify/dsl/each_chain.rb
@@ -5,17 +5,23 @@ require "interactify/dsl/unique_klass_name"
 module Interactify
   module Dsl
     class EachChain
-      attr_reader :each_loop_klasses, :plural_resource_name, :evaluating_receiver
+      MissingIteratableValueInContext = Class.new(ArgumentError)
 
-      def self.attach_klass(evaluating_receiver, plural_resource_name, *each_loop_klasses)
-        iteratable = new(each_loop_klasses, plural_resource_name, evaluating_receiver)
+      attr_reader :each_loop_klasses, :plural_resource_name, :evaluating_receiver, :caller_info
+
+      def self.attach_klass(evaluating_receiver,
+                            *each_loop_klasses,
+                            plural_resource_name:,
+                            caller_info:)
+        iteratable = new(each_loop_klasses, plural_resource_name, evaluating_receiver, caller_info:)
         iteratable.attach_klass
       end
 
-      def initialize(each_loop_klasses, plural_resource_name, evaluating_receiver)
+      def initialize(each_loop_klasses, plural_resource_name, evaluating_receiver, caller_info:)
         @each_loop_klasses = each_loop_klasses
         @plural_resource_name = plural_resource_name
         @evaluating_receiver = evaluating_receiver
+        @caller_info = caller_info
       end
 
       # allows us to dynamically create an interactor chain
@@ -33,10 +39,17 @@ module Interactify
           end                                                                           #   end
 
           define_singleton_method(:source_location) do                                  #   def self.source_location
-            const_source_location this.evaluating_receiver.to_s                         #     [file, line]
+            file, line = this.caller_info&.split(':') 
+            return const_source_location(this.evaluating_receiver.to_s) if file.nil?
+
+            [file, line&.to_i]
           end                                                                           #   end
 
           define_method(:run!) do                                                       #  def run!
+            resources = context.send(this.plural_resource_name)                         #    packages = context.packages
+
+            bail_with_error(resources) unless resources.respond_to?(:each_with_index)   #    raise MissingIteratableValueInContext unless packages.respond_to?(:each_with_index)
+
             context.send(this.plural_resource_name).each_with_index do |resource, index|#    context.packages.each_with_index do |package, index|
               context[this.singular_resource_name] = resource                           #       context.package = package
               context[this.singular_resource_index_name] = index                        #       context.package_index = index
@@ -51,6 +64,11 @@ module Interactify
 
             context                                                                     #     context
           end                                                                           #   end
+
+          define_method(:bail_with_error) do |resources|
+            message = "Expected `context.#{this.plural_resource_name}`: #{resources.inspect}\nto respond to :each_with_index"
+            raise MissingIteratableValueInContext, message
+          end
 
           define_singleton_method(:klasses) do                                          #   def self.klasses
             klasses = instance_variable_get(:@klasses)                                  #     @klasses ||= Wrapper.wrap_many(self, [A, B, C])