+++ spec/fixtures/integration_app/app/interactors/organizing/outer_organizer.rb
@@ -1,0 +1,19 @@
+require_relative 'inner_organizer'
+
+module Organizing
+  class OuterOrganizer
+    include Interactify
+
+    organize \
+      InnerOrganizer.organizing(
+        Organized1, 
+        Organized2.organizing(
+          DeeplyNestedInteractor,
+          DeeplyNestedPromisingInteractor.promising(
+            :deeply_nested_promising_interactor_called
+          ),
+          Organized2::Organized2Called
+        )
+      )
+  end
+end