+++ spec/integration/all_the_things_integration_spec.rb
@@ -14,12 +14,12 @@ RSpec.describe "Interactify" do
   end
 
   before do
-    require_files("each/")
-    require_files("if/")
-    require_files('')
+    load_interactify_fixtures("each")
+    load_interactify_fixtures("if/")
+    load "./spec/fixtures/integration_app/app/interactors/all_the_things.rb"
   end
 
-  context 'without an optional thing' do
+  context "without an optional thing" do
     let(:result) { AllTheThings.promising(:a).call!(things:, optional_thing: false) }
 
     it "sets A and B, then lambda_set, then both_a_and_b, then first_more_thing, next_more_thing" do
@@ -48,14 +48,6 @@ RSpec.describe "Interactify" do
 
       expect(result.counter).to eq 8
       expect(result.heavily_nested_counter).to eq 256
-    end
-  end
-
-  def require_files(dir)
-    files = Dir.glob("./spec/fixtures/integration_app/app/interactors/#{dir}**/*.rb")
-
-    files.each do |file|
-      require file
     end
   end
 end