+++ CHANGELOG.md
@@ -28,3 +28,10 @@
 
 ## [0.4.1] - 2023-12-29
 - Fix bug triggered when nesting each and if
+
+## [0.5.0] - 2024-01-01
+- Add support for `SetA = Interactify { _1.a = 'a' }`, lambda and block class creation syntax
+- Add support for organizing `organize A.organizing(B, C, D), E, F` contract syntax
+- make definition errors raise optionally
+- raise an error with unexpected keys in Interactify.if clause
+- propagate caller_info through chains